requirements.txt
# WBSEDCL Tracking System - Full Requirements (All features)
# Install this for complete functionality including optional features

# ============ CORE PACKAGES ============
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1

# ============ SECURITY ============
bcrypt==4.1.2
passlib==1.7.4

# ============ UTILITIES ============
python-dotenv==1.0.0
python-dateutil==2.8.2
email-validator==2.1.0

# ============ FILE HANDLING ============
Pillow==10.1.0

# ============ EMAIL NOTIFICATIONS ============
Flask-Mail==0.9.1

# ============ DOCUMENT GENERATION ============
# PDF Export
reportlab==4.0.9

# Excel Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# ============ API DEVELOPMENT ============
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# ============ SESSION MANAGEMENT ============
Flask-Session==0.5.0

# ============ QR CODE / BARCODE ============
qrcode==7.4.2
python-barcode==0.15.1

# ============ DEVELOPMENT TOOLS ============
python-decouple==3.8

# ============ TESTING ============
pytest==7.4.3
pytest-flask==1.3.0

# ============ CODE QUALITY ============
flake8==7.0.0
black==23.12.1

# ============ PRODUCTION SERVERS ============
# Windows
waitress==2.1.2

# Linux (uncomment if deploying on Linux)
# gunicorn==21.2.0